+++ db/schema/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS links (
 
 CREATE TABLE IF NOT EXISTS link_visits (
 	id BIGSERIAL PRIMARY KEY,
-	link_id BIGSERIAL,
+	link_id INT,
 	ip VARCHAR(45),
 	user_agent VARCHAR(255),
 	referer TEXT,
